+++ jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/AbstractRingBufferBlockingQueueV2.java
@@ -130,62 +130,4 @@ public abstract class AbstractRingBufferBlockingQueueV2<E> extends AbstractBlock
   public final boolean isEmpty() {
     return head.get() == tail.get();
   }
-
-  @Override
-  public final boolean isFull() {
-    final int queueStart = tail.get() - capacity;
-    return head.get() == queueStart;
-  }
-
-  /**
-   * This is a method description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return Iterator E e.
-   */
-  @Override
-  public Iterator<E> iterator() {
-    throw new UnsupportedOperationException("未实现.");
-  }
-
-  /**
-   * This is a method description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return int e.
-   */
-  @Override
-  public int remainingCapacity() {
-    throw new UnsupportedOperationException("未实现.");
-  }
-
-  /**
-   * This is a method description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return int e.
-   */
-  @Override
-  public int drainTo(final Collection<? super E> c) {
-    throw new UnsupportedOperationException("未实现.");
-  }
-
-  /**
-   * This is a method description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return int e.
-   */
-  @Override
-  public int drainTo(final Collection<? super E> c, final int maxElements) {
-    throw new UnsupportedOperationException("未实现.");
-  }
 }
