+++ jdkringbuffer-jmh/src/main/java/org/jdkstack/jdkringbuffer/jmh/lmax/util/Constants.java
@@ -1,6 +1,17 @@
 package org.jdkstack.jdkringbuffer.jmh.lmax.util;
 
+/**
+ * This is a class description.
+ *
+ * <p>Another description after blank line.
+ *
+ * @author admin
+ */
 public class Constants {
   public static final int RINGBUFFER_SIZE = 1 << 20;
   public static final int ITERATIONS = 1_000_000;
+
+  private Constants() {
+    //
+  }
 }
